+++ src/ui/UIController.ts
@@ -584,7 +584,7 @@ export class UIController {
                         config.get('defaultView'),
                     'scientificDataViewer.allowMultipleTabsForSameFile':
                         config.get('allowMultipleTabsForSameFile'),
-                    'scientificDataViewer.devMode': config.get('devMode'),
+                    'scientificDataViewer.general.devMode': config.get('devMode'),
                     'scientificDataViewer.matplotlibStyle':
                         config.get('matplotlibStyle'),
                 };